Before long to int

+++ src/main/java/com/las/laboratory_appointment_system/model/Doctor.java
@@ -21,7 +21,7 @@ import java.util.List;
 public class Doctor{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long user_id;
+    private int user_id;
     private String username;
     private String password;
     private UserRole role;
