+++ src/main/java/com/las/laboratory_appointment_system/model/Patient.java
@@ -20,7 +20,7 @@ import java.util.List;
 public class Patient{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long user_id;
+    private int user_id;
     private String username;
     private String password;
     private UserRole role;
